requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
motor==3.3.2
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
pytest==7.4.4
python-dotenv==1.0.0
python-multipart==0.0.9
python-jose==3.3.0
PyYAML==6.0.1
ruff==0.3.7
sniffio==1.3.0
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
